Fixing emitTransfer

+++ contracts/IERC721Subordinate.sol
@@ -7,7 +7,6 @@ pragma solidity ^0.8.9;
 // Whoever owns the main token owns the subordinate token.
 // ERC165 interface id is 0x4a5a1d1d
 interface IERC721Subordinate {
-
   // The function dominantToken() returns the address of the dominant token.
   function dominantToken() external view returns (address);
 
@@ -15,5 +14,5 @@ interface IERC721Subordinate {
   // transfer from address 0. This function allow to fix the issue, but
   // it is not mandatory — in same cases, the deployer may want the subordinate
   // being not visible on marketplaces.
-  function emitTransfer(uint tokenId) external;
+  function emitTransfer(uint256 tokenId) external;
 }
